+++ engine/src/flutter/shell/platform/android/platform_view_android_unittests.cc
@@ -29,5 +29,13 @@ TEST(AndroidPlatformView, DISABLED_SelectsVulkanBasedOnApiLevel) {
 }
 #endif  // !SLIMPELLER
 
+TEST(AndroidPlatformView, FallsBackToGLESonMostExynos) {
+  std::string exynos_board = "exynos7870";
+  std::string snap_board = "smg1234";
+
+  EXPECT_TRUE(FlutterMain::IsKnownBadSOC(exynos_board));
+  EXPECT_FALSE(FlutterMain::IsKnownBadSOC(snap_board));
+}
+
 }  // namespace testing
 }  // namespace flutter